Modularized currency box to add to mobile as well
