+++ src/components/Header/styles/blocks/headerMobile.scss
@@ -42,6 +42,10 @@
     }
   }
 
+  &__currency {
+    margin-bottom: 20px;
+  }
+
   &__logo {
     width: 45px;
     height: auto;
